Add :redis/allowed-commands to restrict allowed values for :op
If this whitelist is not defined in the catalog, all operations
supported by carmine are allowed.

The map of all supported operations to their carmine functions
is pre-filtered when creating the writer, so that incoming
operations are looked up only from the allowed commands.

This commit also adds an optional :redis/allowed-commands key
to the writer task spec.